requirements.txt
streamlit
pandas
openpyxl
xlrd==1.2.0
azure-ai-documentintelligence
azure-core
openai
Pillow
